Update Dx project config to be the same than Jack
- Fix all warnings or errors related to the new config.

+++ dx/src/com/android/jack/dx/dex/TableOfContents.java
@@ -250,7 +250,8 @@ public final class TableOfContents {
 
     @Override
     public String toString() {
-      return String.format("Section[type=%#x,off=%#x,size=%#x]", type, off, size);
+      return String.format("Section[type=%#x,off=%#x,size=%#x]", Short.valueOf(type),
+          Integer.valueOf(off), Integer.valueOf(size));
     }
   }
 }
